fix small bug in wizard header title for calibrate
